Input trench morphology
Changed zs, dzsx & dzsy to reflect trench geometry using continuous
functions w/ tanh(x) components. Surface is now parameterized with A,
bx, and by. Still need to fix so that an A>1 runs. Then geometry changes
will be complete and second half of scripts can be modified to accept A,
bx, & by as inputs.